クラス改修
GameScenをMasterSceneとに分割した。
